+++ CHANGELOG.md
@@ -4,7 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+
 - Feature: Add new `InformationBox` React component.
+- Feature: Add click handler on `CallToActionPopover` buttons.
 - Feature: Add new `TextInputWithLimit` React component.
 - Fix: Bring back `@include k-Tooltip;` in _components.scss.